app-backend/requirements/base.txt
bcrypt==3.1.6
boto3==1.9.91
cryptography
https://github.com/illumidesk/django/archive/2.1-uuid-pk.zip
channels==1.1.8
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
docker==2.5.1
drf-nested-routers==0.91
GitPython==2.1.11
hiredis==0.2.0
humanize==0.5.1
nameparser==0.5.7
passlib==1.7.1
psycopg2==2.7.5
Pygments==2.2.0
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
rauth==0.7.3
drf-ujson==1.2
ujson==1.35
django-redis==4.10.0
redis==3.2.0
django-extensions==2.1.6
gunicorn==19.9.0
celery==4.2.1
django-celery-results==1.0.1
uuid==1.30
django-cacheops==4.1
django-storages==1.7.1
django-oauth-toolkit==1.2.0
git+https://github.com/illumidesk/django-rest-framework-social-oauth2.git@1.0.0
djangorestframework-jwt==1.11.0
slackclient==1.2.1
django-cors-headers==2.5.0
django-guardian==1.5.0
Pillow==5.2.0
stripe==1.55.2
django-haystack-elasticsearch==0.1.0
elasticsearch>=5.0.0,<6.0.0
drf-haystack==1.8.4
djoser==1.5.1
django-ses==0.8.6
google-compute-engine==2.8.3
raven==6.9.0
celery-redbeat==0.11.1
django-treebeard==4.3
asgi-redis==1.4.3
django-environ==0.4.5